Move context constant
