+++ src/constants.js
@@ -1,0 +1,2 @@
+// eslint-disable-next-line import/prefer-default-export
+export const CONTEXT_KEY = '__react-amphtml-ampscripts__';
